Refactor area selection into RunPage
