+++ src/components/EncounterTable/EncounterTable.module.scss
@@ -1,17 +1,7 @@
 @use "@/styles/constants";
 
 .encounter-table {
-    --component-spacing: 2rem;
-    --inner-spacing: 1rem;
-    --name-size: 1.7rem;
     --sprite-size: 120px;
-
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-    gap: var(--inner-spacing);
-    justify-content: center;
-    margin: var(--component-spacing) 0;
 
     .header {
         font-size: 1.7rem;
