+++ src/components/EncounterTable/EncounterTable.tsx
@@ -1,4 +1,3 @@
-import Dropdown from "@/components/Dropdown/Dropdown";
 import AreaData from "@/models/AreaData";
 import EncounterData from "@/models/EncounterData";
 import PokemonData from "@/models/PokemonData";
@@ -10,15 +9,13 @@ import styles from "./EncounterTable.module.scss";
 
 type Props = {
     runName: string;
-    areaList: AreaData[];
+    currentArea: AreaData | null;
     starterSlugsList: string[];
     gameGroup: string;
     onFetch: (pokemonDataList: PokemonData[]) => void;
 };
 
 const EncounterTable: React.FC<Props> = (props: Props) => {
-    const [areaNameList, setAreaNameList] = useState<string[]>([]);
-    const [currentArea, setCurrentArea] = useState<AreaData | null>(null);
     const [pokemonDataList, setPokemonDataList] = useState<PokemonData[]>([]);
 
     // Formats the level range of an encounter
@@ -30,18 +27,12 @@ const EncounterTable: React.FC<Props> = (props: Props) => {
         }
     };
 
-    // Sets the current area on dropdown select
-    const handleAreaSelect = (areaName: string) => {
-        const area: AreaData = props.areaList.filter((area: AreaData) => area.areaName === areaName)[0];
-        setCurrentArea(area);
-    };
-
     // API call to fetch Pokemon names and sprites
     const fetchPokemonData = async () => {
-        currentArea!.encounters = currentArea!.encounters.filter(
+        props.currentArea!.encounters = props.currentArea!.encounters.filter(
             (encounter: EncounterData) => !props.starterSlugsList.includes(encounter.pokemonSlug)
         );
-        const pokemonSlugList: string[] = currentArea!.encounters.map(
+        const pokemonSlugList: string[] = props.currentArea!.encounters.map(
             (encounter: EncounterData) => encounter.pokemonSlug
         );
         axios
@@ -62,26 +53,14 @@ const EncounterTable: React.FC<Props> = (props: Props) => {
 
     // Fetch Pokemon names/sprites when area is changed
     useEffect(() => {
-        if (currentArea) {
+        if (props.currentArea) {
             fetchPokemonData();
         }
-    }, [currentArea]);
-
-    // Nullify area when location is changed
-    useEffect(() => {
-        setCurrentArea(null);
-        setAreaNameList(props.areaList.map((area: AreaData) => area.areaName).sort());
-    }, [props.areaList]);
+    }, [props.currentArea]);
 
     return (
         <div className={styles["encounter-table"]}>
-            <h3 className={styles.header}>Encounters:</h3>
-            <Dropdown
-                placeholder="Select a zone..."
-                options={areaNameList}
-                onSelect={(areaName: string) => handleAreaSelect(areaName)}
-            />
-            {currentArea && pokemonDataList.length > 0 ? (
+            {props.currentArea && pokemonDataList.length > 0 ? (
                 <table className={styles.table}>
                     <thead>
                         <tr>
@@ -93,7 +72,7 @@ const EncounterTable: React.FC<Props> = (props: Props) => {
                         </tr>
                     </thead>
                     <tbody>
-                        {currentArea!.encounters.map((encounter: EncounterData, key: number) => {
+                        {props.currentArea.encounters.map((encounter: EncounterData, key: number) => {
                             {
                                 return pokemonDataList[key] ? (
                                     <tr className={styles.row} key={key}>
