+++ src/components/RunPage/RunPage.module.scss
@@ -22,6 +22,18 @@
             letter-spacing: 2px;
             text-align: center;
         }
+
+        .section {
+            align-items: center;
+            display: flex;
+            flex-direction: column;
+            gap: 1rem;
+            margin: 2rem 0;
+
+            .header {
+                font-size: 1.7rem;
+            }
+        }
     }
 
     .sticky-info {
