+++ src/components/StarterSelect/StarterSelect.module.scss
@@ -1,7 +1,6 @@
 @use "@/styles/constants";
 
 .starter-select {
-    --component-spacing: 2rem;
     --name-size: 1.7rem;
     --sprite-size: 120px;
     --inner-spacing: 1rem;
@@ -11,7 +10,6 @@
     flex-direction: column;
     gap: var(--inner-spacing);
     justify-content: center;
-    margin: var(--component-spacing) 0;
 
     .header {
         font-size: 1.7rem;
